refactor(app): move settings lib helpers to feature lib layer
